+++ ice/src/main/java/com/altinity/ice/cli/internal/iceberg/Sorting.java
@@ -25,6 +25,7 @@ import org.apache.iceberg.data.parquet.GenericParquetReaders;
 import org.apache.iceberg.io.CloseableIterable;
 import org.apache.iceberg.io.CloseableIterator;
 import org.apache.iceberg.io.InputFile;
+import org.apache.iceberg.mapping.NameMapping;
 import org.apache.iceberg.parquet.Parquet;
 import org.apache.iceberg.types.Types;
 
@@ -99,14 +100,16 @@ public final class Sorting {
     }
   }
 
-  public static boolean isSorted(InputFile inputFile, Schema tableSchema, SortOrder sortOrder)
+  public static boolean isSorted(
+      InputFile inputFile, Schema tableSchema, SortOrder sortOrder)
       throws IOException {
     return checkSorted(inputFile, tableSchema, sortOrder).ok;
   }
 
   // TODO: check metadata first to avoid full scan when unsorted
   public static SortCheckResult checkSorted(
-      InputFile inputFile, Schema tableSchema, SortOrder sortOrder) throws IOException {
+      InputFile inputFile, Schema tableSchema, SortOrder sortOrder)
+      throws IOException {
     if (sortOrder.isUnsorted()) {
       return new SortCheckResult(false);
     }
@@ -137,7 +140,7 @@ public final class Sorting {
 
     try (CloseableIterable<Record> records =
         Parquet.read(inputFile)
-            .createReaderFunc(s -> GenericParquetReaders.buildReader(projectedSchema, s))
+            .createReaderFunc(s -> GenericParquetReaders.buildReader(tableSchema, s))
             .project(projectedSchema)
             .build()) {